requirements.txt
arch==7.1.0
causal-conv1d @ file:///home/shensi11/installs/mamba/causal_conv1d-1.2.0.post1%2Bcu118torch2.2cxx11abiFALSE-cp312-cp312-linux_x86_64.whl#sha256=24f2d9ace719a7997faba5ffb8f876c3bf564da0ac59717d8f11b773e0072aab
certifi==2024.8.30
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.3.0
cycler==0.12.1
einops==0.8.0
et-xmlfile==1.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
huggingface-hub==0.25.1
hurst==0.0.5
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.43.0
mamba-ssm @ file:///home/shensi11/installs/mamba/mamba_ssm-1.2.0%2Bcu118torch2.2cxx11abiFALSE-cp312-cp312-linux_x86_64.whl#sha256=790af3b323cffceeec0089c6ef431acc50f55b9a8261168f2e6115bd18e367ac
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pillow==10.4.0
psutil==6.1.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
statsmodels==0.14.3
stumpy==1.13.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.0
torch @ file:///home/shensi11/installs/torch-2.2.1%2Bcu118-cp312-cp312-linux_x86_64.whl#sha256=3edd204c8536fb753cffa3e684fd971297c093b63520395ec04c1b88bc8664d0
torchaudio @ file:///home/shensi11/installs/torchaudio-2.2.1%2Bcu118-cp312-cp312-linux_x86_64.whl#sha256=70200e1d223804ad4bde87ff5d86fb2b4e2b3efb8d90e18479c9104f38d22f05
torchinfo==1.8.0
tqdm==4.66.5
transformers==4.45.0
triton==3.0.0
tsfresh==0.20.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.44.0
xlrd==2.0.1
xlstm==1.0.5
